+++ lib/logger.js
@@ -20,7 +20,7 @@ class Logger {
         this._config = config;
         this._serviceName = containerName;
         this._containerName = containerName;
-        this._enableColors = config.enableColors === undefined ? true : config.enableColors;
+        this._enableColors = config.enableColors ?? true;
         if (config.format) {
             this._printFormat = stringTemplate(config.format); // override current format
         }
@@ -132,6 +132,13 @@ class Logger {
                     onConfigChange: level => {
                         this._config.verbosityLevel = level;
                     }
+                })
+            );
+        }
+        if (config.transport.file) {
+            transports.push(
+                new winston.transports.File({
+                    ...config.fileTransport
                 })
             );
         }
