+++ lib/plugins/verbosity.js
@@ -1,12 +1,8 @@
-/*
- * Created by nassi on 02/02/17.
- */
-
 const RedisFactory = require('@hkube/redis-utils').Factory;
 const PubSubAdapter = require('@hkube/pub-sub-adapter');
 const BasePlugin = require('./basePlugin');
 const levels = require('../consts/levels');
-const TOPIC = 'rms-logger-api-trace-level';
+const TOPIC = 'logger-api-trace-level';
 let uniqueKey;
 let client;
