+++ package.json
@@ -12,9 +12,9 @@
 		"stack-trace": "0.0.9",
 		"string-template": "^1.0.0",
 		"uuid": "^8.1.0",
-		"winston": "^3.2.1",
-		"winston-logstash": "^0.2.11",
-		"winston-transport": "^4.3.0"
+		"winston": "^3.3.3",
+		"winston-logstash": "^0.4.0",
+		"winston-transport": "^4.4.0"
 	},
 	"devDependencies": {
 		"chai": "^3.3.0",
